Add LumpManager class to make file searches quicker
Repeatedly using glob.glob or glob.iglob takes a lot of time, since the filesystem has to be accessed and queried each time either function is called. It's much quicker to look up values from a dictionary which was created and populated by calling glob.iglob than it is to call glob.glob or glob.iglob repeatedly.